+++ src/gc.c
@@ -73,7 +73,7 @@ gc_copy (SCM old) ///((internal))
   if (NTYPE (new) == TVECTOR)
     {
       NVECTOR (new) = g_free;
-      for (int i=0; i<LENGTH (old); i++)
+      for (long i=0; i<LENGTH (old); i++)
         g_news[g_free++] = g_cells[VECTOR (old)+i];
     }
   TYPE (old) = TBROKEN_HEART;
@@ -172,7 +172,7 @@ gc_ () ///((internal))
   g_free = 1;
 
 #if __MESC__
-  if (ARENA_SIZE < MAX_ARENA_SIZE && (int)g_news > 0)
+  if (ARENA_SIZE < MAX_ARENA_SIZE && (long)g_news > 0)
 #else
   if (ARENA_SIZE < MAX_ARENA_SIZE && g_news > 0)
 #endif
@@ -194,7 +194,7 @@ gc_ () ///((internal))
       gc_up_arena ();
     }
 
-  for (int i=g_free; i<g_symbol_max; i++)
+  for (long i=g_free; i<g_symbol_max; i++)
     gc_copy (i);
   g_symbols = gc_copy (g_symbols);
   g_macros = gc_copy (g_macros);
